Made conditional template directories more obvious

+++ Mac/clean.command
@@ -10,9 +10,11 @@ export projectHomeDir=`dirname "$scriptDir"`
 
 cd "$projectHomeDir"
 
-export EXTENSION_DIRNAME=`head -n 1 "$projectHomeDir/BuildSettings/ExtensionDirName.txt"`
-export EXTENSION_HOMEDIR=~/Library/Application\ Support/Adobe/CEP/extensions/"$EXTENSION_DIRNAME"
+if [ -f "$projectHomeDir/BuildSettings/ExtensionDirName.txt" ]; then
+	export EXTENSION_DIRNAME=`head -n 1 "$projectHomeDir/BuildSettings/ExtensionDirName.txt"`
+	export EXTENSION_HOMEDIR=~/Library/Application\ Support/Adobe/CEP/extensions/"$EXTENSION_DIRNAME"
+	rm -rf "$EXTENSION_HOMEDIR"
+fi
 
 rm -rf LocalLinks
 rm -rf build
-rm -rf "$EXTENSION_HOMEDIR"
